fixed bug #1565, now u can't add or edit course,online course, study, grant,council with duplicate topics. fixed bug #1572, you can edit only your events. fixed bug #1501, you can see only approved courses and tutors

+++ PortalLib/src/main/java/ua/dp/stud/StudPortalLib/dao/CourseDao.java
@@ -39,5 +39,5 @@ public interface CourseDao extends DaoForApprove<Course>{
     void initializeCountOfCourses(KindOfCourse kindOfCourse);
 
 
-    Boolean isDuplicateTopic(String name, Long id);
+    Boolean isDuplicateTopic(String name,Integer id);
 }
