+++ PortalLib/src/main/java/ua/dp/stud/StudPortalLib/service/CourseService.java
@@ -49,6 +49,6 @@ public interface CourseService {
 
     void initializeCountOfCourses(KindOfCourse kindOfCourse);
 
-    Boolean isDuplicateTopic(String name,Long id);
+    Boolean isDuplicateTopic(String name,Integer id);
 
 }
