+++ PortalLib/src/main/java/ua/dp/stud/StudPortalLib/service/impl/CourseServiceImpl.java
@@ -145,7 +145,7 @@ public class CourseServiceImpl implements CourseService {
     }
 
     @Override
-    public Boolean isDuplicateTopic(String name,Long id){
+    public Boolean isDuplicateTopic(String name,Integer id){
         return dao.isDuplicateTopic(name,id);
     }
 }
