+++ Studie/src/main/java/ua/dp/stud/studie/dao/StudieDao.java
@@ -19,6 +19,6 @@ public interface StudieDao extends DaoForApprove<Studie> {
 
     Collection<Studie> getAll();
 
-    Boolean isDuplicateTopic(String name, Long id);
+    Boolean isDuplicateTopic(String name, Integer id);
 
 }
