+++ Studie/src/main/java/ua/dp/stud/studie/service/GrantService.java
@@ -19,5 +19,5 @@ public interface GrantService {
 
     void updateGrant(Grant grant);
 
-    Boolean isDuplicateTopic(String name, Long id);
+    Boolean isDuplicateTopic(String name, Integer id);
 }
