+++ Studie/src/main/java/ua/dp/stud/studie/service/impl/GrantServiceImpl.java
@@ -56,7 +56,7 @@ public class GrantServiceImpl implements GrantService {
 	}
 
     @Override
-    public Boolean isDuplicateTopic(String name, Long id){
+    public Boolean isDuplicateTopic(String name, Integer id){
         return dao.isDuplicateTopic(name,id);
     }
