+++ Studie/src/main/java/ua/dp/stud/studie/service/impl/OnlineCourseServiceImpl.java
@@ -139,7 +139,7 @@ public class OnlineCourseServiceImpl implements OnlineCourseService
     }
 
     @Override
-    public Boolean isDuplicateTopic(String name, Long id){
+    public Boolean isDuplicateTopic(String name, Integer id){
         return dao.isDuplicateTopic(name,id);
     }
